Because we updated the UserAvatar to be dynamic, rather than separate components, we need to update its references elsewhere.
