+++ app/src/components/common/UserAvatar.tsx
