+++ app/src/components/proposals/proposalPage/VotingSidebar/votingColumn/CandidateCard.tsx
@@ -2,7 +2,7 @@
 
 import { User } from "@prisma/client"
 
-import { UserAvatar } from "@/components/common/UserAvatarLarge"
+import { UserAvatar } from "@/components/common/UserAvatar"
 import { useUserProjects } from "@/hooks/db/useUserProjects"
 import { ProjectWithDetails } from "@/lib/types"
